refactor according to comments
