+++ src/components/Root.js
@@ -6,7 +6,7 @@ import useGoogleAnalytics from 'hooks/useGoogleAnalytics';
 import usePixel from 'hooks/usePixel';
 import useGoogleTagManager from 'hooks/useGoogleTagManager';
 import useSyncAuth from 'hooks/useSyncAuth';
-import useAmplitude from 'hooks/useAmplitude';
+import useAmplitudeLandPage from 'hooks/useAmplitudeLandPage';
 import App from './App';
 
 const Root = () => {
@@ -14,7 +14,7 @@ const Root = () => {
   usePixel();
   useGoogleTagManager();
   useSyncAuth();
-  useAmplitude();
+  useAmplitudeLandPage();
 
   return (
     <PermissionContextProvider>
