+++ src/hooks/useAmplitudeLandPage.js
